Drop the extra cloneInto experimentation.
There was more experimentation on wrapping than actually was necessary
because the problem was an un-wrapped/cloned "this" handed into the
content global by the now-moot `_register` method.  But this had not
been immediately recognized as the problem because initially there were
problems in the wrapping logic due to not traversing the prototype
chain.