Caching seems to work, task-tree fixed up.
The experience of using and iterating on the analyzer was bad as the
latency of asking pernosco the same things over and over approached
thirty seconds so I added a cache in IndexedDB.  There currently isn't
any eviction/cleanup of the cache but an index is in place to enable it.

The task-tree broke when running against an older trace where the
assumption about the placement of the apparent `exec` invocations were
disrupted (there was a short-lived fork preceding it), so things are
now simplified with exec treated as a distinct operation.